removed unnecessary methods and reviewed documentation
